// src/app/shared/models/configMatriz.ts
export interface ConfigMatriz{
    idConfigMatriz?: string;
    tipoMatriz?: string;
    nivel?: string;
    proceso?: string;
    coConfigMatriz?: string;
    cantidadRiesgos?: string;
    dimensionAtc?: string;
    tipoAtc?: string;
    periodo?: string;
    nombreMatriz?: string;
    usuarioCreacion?: string;
    usuarioModificacion?: string;
}